skin-disease-ai/requirements.txt
tensorflow-cpu>=2.10.0
flask>=2.3.0
flask-cors>=4.0.0
opencv-python-headless>=4.8.0
numpy>=1.23.0
pandas>=1.5.0
Pillow>=9.4.0
matplotlib>=3.6.0
scikit-learn>=1.2.0
gunicorn>=21.2.0
